#122 Update Flow: Report issue on Camera

+++ src/com/cloudjay/cjay/AuditorContainerActivity.java
@@ -1,7 +1,6 @@
 package com.cloudjay.cjay;
 
 import java.sql.SQLException;
-import java.util.UUID;
 
 import org.androidannotations.annotations.AfterViews;
 import org.androidannotations.annotations.Background;
@@ -12,20 +11,13 @@ import org.androidannotations.annotations.ItemClick;
 import org.androidannotations.annotations.ItemLongClick;
 import org.androidannotations.annotations.OptionsItem;
 import org.androidannotations.annotations.OptionsMenu;
-import org.androidannotations.annotations.Trace;
 import org.androidannotations.annotations.UiThread;
 import org.androidannotations.annotations.ViewById;
 
-import android.app.AlertDialog;
 import android.content.Context;
-import android.content.DialogInterface;
-import android.content.DialogInterface.OnClickListener;
-import android.content.Intent;
 import android.database.Cursor;
-import android.database.sqlite.SQLiteDatabase;
 import android.os.Bundle;
 import android.text.TextUtils;
-import android.util.Log;
 import android.widget.ImageButton;
 import android.widget.ListView;
 import android.widget.TextView;
@@ -38,15 +30,14 @@ import com.cloudjay.cjay.dao.ContainerSessionDaoImpl;
 import com.cloudjay.cjay.dao.DamageCodeDaoImpl;
 import com.cloudjay.cjay.dao.IssueDaoImpl;
 import com.cloudjay.cjay.dao.RepairCodeDaoImpl;
-import com.cloudjay.cjay.events.CJayImageAddedEvent;
 import com.cloudjay.cjay.model.CJayImage;
 import com.cloudjay.cjay.model.ContainerSession;
 import com.cloudjay.cjay.model.Issue;
 import com.cloudjay.cjay.util.CJayConstant;
 import com.cloudjay.cjay.util.CJayCustomCursorLoader;
 import com.cloudjay.cjay.util.DataCenter;
+import com.cloudjay.cjay.util.IssueReportHelper;
 import com.cloudjay.cjay.util.Logger;
-import com.cloudjay.cjay.util.Utils;
 
 import de.keyboardsurfer.android.widget.crouton.Crouton;
 import de.keyboardsurfer.android.widget.crouton.Style;
@@ -184,9 +175,9 @@ public class AuditorContainerActivity extends CJayActivity implements android.ap
 		Logger.e("Click on " + issueId);
 
 		if (!TextUtils.isEmpty(issueId)) {
-			showIssueReport(mSelectedCJayImageUuid);
+			IssueReportHelper.showIssueReport(this, mSelectedCJayImageUuid);
 		} else {
-			showReportDialog();
+			IssueReportHelper.showReportDialog(this, mSelectedCJayImageUuid, mContainerSessionUUID);
 		}
 
 	}
@@ -203,6 +194,10 @@ public class AuditorContainerActivity extends CJayActivity implements android.ap
 
 		try {
 			mLongClickedCJayImage = cJayImageDaoImpl.findByUuid(mLongClickedCJayImageUuid);
+			Issue issue = mLongClickedCJayImage.getIssue();
+			if (issue != null) {
+				issue.equals(issue);
+			}
 		} catch (SQLException e) {
 
 			e.printStackTrace();
@@ -232,25 +227,6 @@ public class AuditorContainerActivity extends CJayActivity implements android.ap
 				return cursor;
 			}
 		};
-	}
-
-	public void onEvent(CJayImageAddedEvent event) {
-		Logger.Log("issue_report - " + event.getCJayImage().getUuid() + " - cjayimage added - " + event.getTag());
-		if (event.getTag().equals(LOG_TAG)) {
-			mNewImageCount++;
-			mNewImageUUID = event.getCJayImage().getUuid();
-
-			Logger.Log("issue_report - isRunning=" + isRunning());
-			if (isRunning()) {
-				if (mCursorAdapter != null) {
-					Logger.Log("issue_report - refresh");
-					refresh();
-				}
-
-				Logger.Log("issue_report - " + mNewImageUUID + " - start - imageCount=" + mNewImageCount);
-				showReportDialogForNewImage();
-			}
-		}
 	}
 
 	@Override
@@ -300,156 +276,23 @@ public class AuditorContainerActivity extends CJayActivity implements android.ap
 		} else {
 			Logger.Log("issue_report - refresh");
 
-			try {
-				containerSessionDaoImpl.refresh(mContainerSession);
-			} catch (SQLException e) {
-				e.printStackTrace();
-			}
-
 			if (mCursorAdapter != null) {
 				// otherwise refresh the image list
 				refresh();
-			}
-
-			if (mNewImageCount == 1 && !TextUtils.isEmpty(mNewImageUUID)) {
-				// go to report issue after taking one picture
-				showReportDialogForNewImage();
 			}
 		}
 	}
 
 	@UiThread
 	public void refresh() {
-		getLoaderManager().restartLoader(LOADER_ID, null, this);
-		supportInvalidateOptionsMenu();
-	}
-
-	@Background
-	void setWWContainer() {
-		long startTime = System.currentTimeMillis();
-		Logger.Log("*** Create water wash issue***");
-
-		SQLiteDatabase db = DataCenter.getDatabaseHelper(context).getWritableDatabase();
-
-		Cursor damageCursor = db.rawQuery("select id as _id from damage_code where code = ?", new String[] { "DB" });
-
-		int damageId = 0;
-		if (damageCursor.moveToFirst()) {
-			damageId = damageCursor.getInt(damageCursor.getColumnIndexOrThrow("_id"));
-		}
-
-		Cursor repairCursor = db.rawQuery("select id as _id from repair_code where code = ?", new String[] { "WW" });
-
-		int repairId = 0;
-		if (repairCursor.moveToFirst()) {
-			repairId = repairCursor.getInt(repairCursor.getColumnIndexOrThrow("_id"));
-		}
-
-		Cursor componentCursor = db.rawQuery("select id as _id from component_code where code = ?",	new String[] { "FWA" });
-
-		int componentId = 0;
-		if (componentCursor.moveToFirst()) {
-			componentId = componentCursor.getInt(componentCursor.getColumnIndexOrThrow("_id"));
-		}
-		
-		String issueId;
-		String sql = "select _id from issue where componentCode_id = " + componentId + " and damageCode_id = " + damageId
-				+ " and repairCode_id = " + repairId + " and locationCode = ? and containerSession_id = ?";
-		Cursor issueCursor = db.rawQuery(sql, new String[] { "BXXX", mContainerSessionUUID });
-		if (issueCursor.moveToFirst()) {
-			// a WW issue already exists. Update quantity
-			issueId = issueCursor.getString(repairCursor.getColumnIndexOrThrow("_id"));
-			sql = "update issue set quantity = quantity + 1 where _id = " + Utils.sqlString(issueId);
-			db.execSQL(sql);
-		} else {
-			// create a new WW issue
-			issueId = UUID.randomUUID().toString();
-			sql = "insert into issue "
-					+ "(componentCode_id, containerSession_id, damageCode_id, _id, height, repairCode_id, length, locationCode, quantity, id, fixed) "
-					+ " VALUES " + "(" + componentId + ", '" + mContainerSessionUUID + "', " + damageId + ", " + Utils.sqlString(issueId)
-					+ ", NULL, " + repairId + ", NULL, 'BXXX', 1, 0, 0)";
-			db.execSQL(sql);
-		}
-		
-		// link issue to cjayimage
-		sql = "UPDATE cjay_image SET issue_id = '" + issueId + "' WHERE uuid = '" + mSelectedCJayImageUuid + "'";
-		db.execSQL(sql);
-		
 		try {
 			containerSessionDaoImpl.refresh(mContainerSession);
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
-
-		refresh();
-
-		// cost 50ms - Vu: updated cost: ~150-170ms
-		long difference = System.currentTimeMillis() - startTime;
-		Logger.w("---> Total time: " + Long.toString(difference));
-	}
-
-	private void showIssueAssigment(String imageUuid) {
-
-		Intent intent = new Intent(this, AuditorIssueAssigmentActivity_.class);
-		intent.putExtra(AuditorIssueAssigmentActivity.CJAY_IMAGE_EXTRA, imageUuid);
-		startActivity(intent);
-
-	}
-
-	private void showIssueReport(String imageUuid) {
-
-		Intent intent = new Intent(this, AuditorIssueReportActivity_.class);
-		intent.putExtra(AuditorIssueReportActivity.CJAY_IMAGE_EXTRA, imageUuid);
-		startActivity(intent);
-
-	}
-
-	@Trace(level = Log.INFO)
-	void showReportDialog() {
-
-		AlertDialog.Builder builder = new AlertDialog.Builder(this)
-			.setMessage(R.string.dialog_report_message)
-			.setTitle(R.string.dialog_report_title)
-			.setPositiveButton(	R.string.dialog_report_no,
-								new DialogInterface.OnClickListener() {
-									@Override
-									public void onClick(DialogInterface dialog, int id) {
-	
-										// Issue not reported, report issue
-										showIssueReport(mSelectedCJayImageUuid);
-									}
-								})
-			.setNegativeButton(	R.string.dialog_report_yes,
-								new DialogInterface.OnClickListener() {
-									@Override
-									public void onClick(DialogInterface dialog, int id) {
-	
-										// The issue already reported, assign this image to that issue
-										showIssueAssigment(mSelectedCJayImageUuid);
-									}
-								})
-			.setNeutralButton(R.string.dialog_report_neutral,
-								new OnClickListener() {
-									@Override
-									public void onClick(DialogInterface dialog, int which) {
-	
-										setWWContainer();
-									}
-								});
-		builder.show();
-	}
-
-	private void showReportDialogForNewImage() {
-		if (isRunning() && mNewImageCount == 1 && !TextUtils.isEmpty(mNewImageUUID)) {
-			Logger.Log("issue_report");
-
-			// go to report issue after taking one picture
-			mSelectedCJayImageUuid = mNewImageUUID;
-			mNewImageCount = 0;
-			mNewImageUUID = "";
-
-			showReportDialog();
-		}
+		
+		getLoaderManager().restartLoader(LOADER_ID, null, this);
+		supportInvalidateOptionsMenu();
 	}
 
 	@OptionsItem(R.id.menu_trash)
@@ -461,14 +304,13 @@ public class AuditorContainerActivity extends CJayActivity implements android.ap
 
 			// update records in db
 			try {
-
 				cJayImageDaoImpl.delete(mLongClickedCJayImage);
-				containerSessionDaoImpl.refresh(mContainerSession);
-				issueDaoImpl.refresh(issue);
-				if (issue.getCJayImages().size() == 0) {
-					issueDaoImpl.delete(issue);
+				if (issue != null) {
+					issueDaoImpl.refresh(issue);
+					if (issue.getCJayImages().size() == 0) {
+						issueDaoImpl.delete(issue);
+					}					
 				}
-
 			} catch (SQLException e) {
 				e.printStackTrace();
 			}
